Bug 565143 - [Passage] floating licensing
all new API moved to .internal.
