+++ bundles/org.eclipse.passage.lbc.api/src/org/eclipse/passage/lbc/internal/api/LicenseAlreadyTakenException.java
@@ -10,17 +10,17 @@
  * Contributors:
  *     ArSysOp - initial API and implementation
  *******************************************************************************/
-package org.eclipse.passage.lbc.api;
+package org.eclipse.passage.lbc.internal.api;
 
 import org.eclipse.passage.lic.internal.api.conditions.ConditionPack;
 
 /**
  * @since 1.0
  */
-public class NoSuchTakenLicenseException extends Exception {
+public class LicenseAlreadyTakenException extends Exception {
 
-	public NoSuchTakenLicenseException(ConditionPack conditionPack) {
-		super("No such taken license on the server: " + conditionPack.toString()); //$NON-NLS-1$
+	public LicenseAlreadyTakenException(ConditionPack conditionPack) {
+		super("License is already taken: " + conditionPack.toString()); //$NON-NLS-1$
 	}
 
 }
